Snowflake/sfsnippets/invalid-sf-view-identification-master@c52984b3414/requirements.txt
asn1crypto==1.3.0
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
beautifulsoup4==4.9.1
boto3==1.13.19
botocore==1.16.19
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
cryptography==2.9.2
cx-Oracle==7.3.0
docutils==0.15.2
hvac==0.10.3
idna==2.9
ijson==2.6.1
Jinja2==2.11.2
jmespath==0.10.0
markdown2==2.3.9
MarkupSafe==1.1.1
numpy==1.18.4
oscrypto==1.2.0
pandas==1.0.4
pycparser==2.20
pycryptodomex==3.9.7
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
s3transfer==0.3.3
six==1.15.0
snowflake-connector-python==2.2.6
snowflake-sqlalchemy==1.2.3
soupsieve==2.0.1
SQLAlchemy==1.3.17
urllib3==1.25.9
